Move SlayerUnlock to net.runelite.api.vars
