+++ runelite-api/src/main/java/net/runelite/api/vars/SlayerUnlock.java
@@ -22,7 +22,11 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package net.runelite.api;
+package net.runelite.api.vars;
+
+import net.runelite.api.Client;
+import net.runelite.api.VarPlayer;
+import net.runelite.api.Varbits;
 
 /**
  * Unlockables in the slayer interface
