+++ runelite-client/src/main/java/net/runelite/client/plugins/slayer/SlayerPlugin.java
@@ -50,7 +50,7 @@ import net.runelite.api.ItemID;
 import net.runelite.api.NPC;
 import net.runelite.api.NPCComposition;
 import static net.runelite.api.Skill.SLAYER;
-import net.runelite.api.SlayerUnlock;
+import net.runelite.api.vars.SlayerUnlock;
 import net.runelite.api.coords.WorldPoint;
 import net.runelite.api.events.ChatMessage;
 import net.runelite.api.events.ConfigChanged;
